requirements.txt
pandas>=1.1.4
elasticsearch~=7.10.0
numpy~=1.19.5
scikit-learn~=0.24.0
pytest~=6.2.2
pytest_mock~=3.6.1
PyYAML~=5.4.1
fastapi>=0.65.1
uvicorn>=0.14.0
jsonlines>=2.0.0
